requirements.txt
torch
torchvision
transformers
pillow
opencv-python
